ai_pipeline_testing/requirements.txt
openai>=1.40.0
pandas>=2.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
